% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_desc.R
\name{fct_desc}
\alias{fct_desc}
\title{A version of \code{reorder()} function that uses \code{desc()} from \link{dplyr}}
\usage{
fct_desc(x)
}
\arguments{
\item{x, }{a vector of values}
}
\value{
x as a factor, levels reordered in descending order
}
\description{
This function converts the vector of values into a factor
and rearranges levels in a descending order
}
\examples{
fct_desc(c(1, 3, 10, 2, 20))
fct_desc(c("Spring", "Fall", "Summer", "Winter"))
}
